removed extra test end points
